requirements.txt
streamlit
pandas
gspread
google-auth
matplotlib
plotly
pillow
supabase
toml
